feat: Agrega estilos base en CSS y vistas para dashboard

+++ src/Controller/Entertainment/EntertainmentDeleteController.php
@@ -1,7 +1,6 @@
 <?php 
 
 include_once $_SERVER["DOCUMENT_ROOT"].'/src/Controller/SesionController.php';
-use Src\Utils\ControllerUtils;
 use Src\Service\Entertainment\EntertainmentDeleterService;
 
 final readonly class EntertainmentDeleteController extends SesionController {
@@ -14,7 +13,16 @@ final readonly class EntertainmentDeleteController extends SesionController {
     public function start(): void 
     {
         $this->validateUser();
-        $id = $_POST["id"] ?? 0;
-        $this->service->delete($id);
+        if ($_SERVER['REQUEST_METHOD'] === 'POST') {
+            $id = $_POST['id'] ?? null;
+
+            if ($id !== null) {
+                $this->service->delete($id);
+                echo json_encode(['success' => true]);
+            } else {
+                echo json_encode(['success' => false, 'message' => 'ID no enviado']);
+            }
+        }
+        
     }
 }
